post with tags connection and http session for post

+++ src/main/java/com/ittalens/gag/controller/PostController.java
@@ -2,11 +2,13 @@ package com.ittalens.gag.controller;
 
 import com.ittalens.gag.model.dto.posts.PostCreateReqDto;
 import com.ittalens.gag.model.dto.posts.PostRespDto;
+import com.ittalens.gag.model.exceptions.UnauthorizedException;
 import com.ittalens.gag.services.PostServiceImpl;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 @RestController
@@ -16,30 +18,45 @@ public class PostController {
     private final PostServiceImpl postService;
 
     @PostMapping("/upload")
-    public ResponseEntity<?> uploadFile(@ModelAttribute PostCreateReqDto dto) {
-        postService.createPost(dto);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<?> uploadFile(@ModelAttribute PostCreateReqDto dto, HttpSession session) {
+        if (session.getAttribute("LOGGED") != null) {
+            postService.createPost(dto);
+            return ResponseEntity.ok().build();
+        }
+        throw new UnauthorizedException("Must to be logged");
     }
 
     @GetMapping("/post/all")
-    public List<PostRespDto> getAllPosts() {
-        return postService.getAllPostsDto();
+    public List<PostRespDto> getAllPosts(HttpSession session) {
+        if (session.getAttribute("LOGGED") != null) {
+            return postService.getAllPostsDto();
+        }
+        throw new UnauthorizedException("Must to be logged");
     }
 
     @GetMapping("/post/date")
-    public List<PostRespDto> getAllPostsByDateAndTime() {
-        return postService.getAllByCreationDate();
+    public List<PostRespDto> getAllPostsByDateAndTime(HttpSession session) {
+        if (session.getAttribute("LOGGED") != null) {
+            return postService.getAllByCreationDate();
+        }
+        throw new UnauthorizedException("Must to be logged");
     }
 
     @GetMapping("/post")
-    public ResponseEntity<?> getAllPostsByWord(@RequestParam String word) {
-        List<PostRespDto> posts = postService.findPostsByWord(word);
-        return ResponseEntity.ok(posts);
+    public ResponseEntity<?> getAllPostsByWord(@RequestParam String word, HttpSession session) {
+        if (session.getAttribute("LOGGED") != null) {
+            List<PostRespDto> posts = postService.findPostsByWord(word);
+            return ResponseEntity.ok(posts);
+        }
+        throw new UnauthorizedException("Must to be logged");
     }
 
     @DeleteMapping("/post")
-    public ResponseEntity<?> deletedPost(@RequestParam Long id) {
-        postService.deletedPostById(id);
-        return ResponseEntity.ok().build();
+    public ResponseEntity<?> deletedPost(@RequestParam Long id, HttpSession session) {
+        if (session.getAttribute("LOGGED") != null) {
+            postService.deletedPostById(id);
+            return ResponseEntity.ok().build();
+        }
+        throw new UnauthorizedException("Must to be logged");
     }
 }
